refactor: 백업 파일 생성 로직 수정

+++ HRBank/src/main/java/org/yebigun/hrbank/domain/binaryContent/storage/BinaryContentStorageImpl.java
@@ -3,6 +3,7 @@ package org.yebigun.hrbank.domain.binaryContent.storage;
 import jakarta.annotation.PostConstruct;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.InputStreamResource;
 import org.springframework.http.HttpHeaders;
@@ -31,7 +32,7 @@ import static org.yebigun.hrbank.domain.binaryContent.entity.QBinaryContent.bina
  * Author       : dounguk
  * Date         : 2025. 6. 5.
  */
-
+@Log4j2
 @Transactional
 @Service
 @RequiredArgsConstructor
@@ -129,36 +130,39 @@ public class BinaryContentStorageImpl implements BinaryContentStorage {
 
     @Override
     public BinaryContent putCsv(List<Employee> employees) {
-        // 1. 유저 전부 불러오고
-        // 2. UUID를 이름으로 파일 만들고
-        //      (1). UUID 생성
         UUID tempFileName = UUID.randomUUID();
         Path tempPath = root.resolve(tempFileName + CSV_EXTENTION);
 
+        log.warn("파일 생성 시작");
         try (BufferedWriter bw = new BufferedWriter(new FileWriter(tempPath.toFile()))) {
-            //      (2). column 추가
+            log.warn("파일 생성");
             bw.write(COLUMNS);
             bw.newLine();
-            //      (3). employee 정보 추가
-            for (Employee employee : employees) {
-                try{
-                    bw.write(String.format("%d,%s,%s,%s,%s,%s,%s,%s",
-                        employee.getId(),
-                        employee.getEmployeeNumber(),
-                        employee.getName(),
-                        employee.getEmail(),
-//                        employee.getDepartment().getId(),
-                        "[employee.getDepartment().getId() <- dto 추가되면 매핑해서 추가]",
-                        employee.getPosition(),
-                        employee.getHireDate(),
-                        employee.getStatus()
-                    ));
-                    bw.newLine();
-                } catch (Exception e){
+
+            if(employees != null || !employees.isEmpty()) {
+                for (Employee employee : employees) {
+                    log.warn("내부 반복");
                     try {
-                        Files.deleteIfExists(tempPath);
-                    } catch (IOException deleteException) {
-                        e.addSuppressed(deleteException);
+                        bw.write(String.format("%d,%s,%s,%s,%s,%s,%s,%s",
+                            employee.getId(),
+                            employee.getEmployeeNumber(),
+                            employee.getName(),
+                            employee.getEmail(),
+//                        employee.getDepartment().getId(),
+                            "[employee.getDepartment().getId() <- dto 추가되면 매핑해서 추가]",
+                            employee.getPosition(),
+                            employee.getHireDate(),
+                            employee.getStatus()
+                        ));
+                        bw.newLine();
+                    } catch (Exception e) {
+                        log.warn("직원 정보 쓰기 실패: " + employee.getId(), e);
+                        try {
+                            Files.deleteIfExists(tempPath);
+                            log.warn("CSV 파일 삭제");
+                        } catch (IOException deleteException) {
+                            e.addSuppressed(deleteException);
+                        }
                     }
                 }
             }
@@ -167,12 +171,11 @@ public class BinaryContentStorageImpl implements BinaryContentStorage {
             throw new RuntimeException("파일 생성중 오류 발생", e);
         }
 
-        // 3. 파일로 바이너리 컨텐츠 만들고
         long fileSize;
         try{
             fileSize = Files.size(tempPath);
         } catch (IOException e) {
-            throw new RuntimeException("파일 저장에 실패했습니다", e);
+            throw new RuntimeException("파일 사이즈 계산에 실패했습니다", e);
         }
 
         BinaryContent binaryContent = BinaryContent.builder()
@@ -231,7 +234,7 @@ public class BinaryContentStorageImpl implements BinaryContentStorage {
         }
 
         BinaryContent binaryContent = BinaryContent.builder()
-            .fileName(tempFileName + LOG_EXTENTION)
+            .fileName(backupId + LOG_EXTENTION)
             .size(fileSize)
             .contentType(LOG_CONTENT_TYPE)
             .build();
